feat: 추천 스텝 컴포넌트 레이어 이동
