+++ frontend/apps/client/src/app/(main)/recommend/_components/index.ts
@@ -1,4 +1,1 @@
 export { default as FlowHeader } from './FlowHeader';
-export { default as StepAnalyze } from './StepAnalyze';
-export { default as StepGenre } from './StepGenre';
-export { default as StepSituation } from './StepSituation';
