+++ frontend/apps/client/src/features/recommend/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as StepAnalyze } from './StepAnalyze';
+export { default as StepSituation } from './StepSituation';
+export { default as StepGenre } from './StepGenre';
